# docs/openapi.yaml
openapi: 3.1.0
info:
  title: FitnessTrack Public API
  version: 1.0.0
  description: >-
    REST API for the FitnessTrack platform. All endpoints respond with JSON and
    leverage RFC 7807 problem-details for errors.
servers:
  - url: /api/v1
    description: Version 1 base path
security:
  - bearerAuth: []
...
paths:
...
  /subjects:
    get:
      tags: [Subjects]
      operationId: listSubjects
      summary: List subjects
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Sort'
        - name: user_id
          in: query
          schema:
            type: integer
        - name: pseudonym
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Subject page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectPage'
    post:
      tags: [Subjects]
      operationId: createSubject
      summary: Create subject
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubjectCreate'
      responses:
        '201':
          description: Created subject
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subject'
  /subjects/{subjectId}:
    parameters:
      - $ref: '#/components/parameters/SubjectId'
    get:
      tags: [Subjects]
      operationId: getSubject
      summary: Retrieve subject
      responses:
        '200':
          description: Subject document
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subject'
        '404':
          $ref: '#/components/responses/Problem'
    patch:
      tags: [Subjects]
      operationId: updateSubject
      summary: Update subject
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubjectUpdate'
      responses:
        '200':
          description: Updated subject
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subject'
        '412':
          $ref: '#/components/responses/Problem'
    delete:
      tags: [Subjects]
      operationId: deleteSubject
      summary: Delete subject
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Deleted
  /subjects/{subjectId}/profile:
    parameters:
      - $ref: '#/components/parameters/SubjectId'
    get:
      tags: [Subjects]
      operationId: getSubjectProfile
      summary: Retrieve subject profile
      responses:
        '200':
          description: Subject profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectProfile'
        '404':
          $ref: '#/components/responses/Problem'
    put:
      tags: [Subjects]
      operationId: upsertSubjectProfile
      summary: Replace subject profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubjectProfileUpdate'
      responses:
        '200':
          description: Profile document
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectProfile'
        '412':
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
      description: Page number (1-indexed)
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
      description: Page size
    Sort:
      name: sort
      in: query
      schema:
        type: string
      description: Comma separated sort fields, prefix with '-' for descending.
...
    SubjectId:
      name: subjectId
      in: path
      required: true
      schema:
        type: integer
...
  responses:
    Problem:
      description: Problem details error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
        errors:
          oneOf:
            - type: object
            - type: array
              items:
                type: object
    PageEnvelope:
      type: object
      properties:
        items:
          type: array
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
      required: [items, page, limit, total]
...
    Subject:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
          nullable: true
        pseudonym:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SubjectCreate:
      type: object
      properties:
        user_id:
          type: integer
          nullable: true
    SubjectUpdate:
      type: object
      properties:
        user_id:
          type: integer
          nullable: true
    SubjectPage:
      allOf:
        - $ref: '#/components/schemas/PageEnvelope'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Subject'
    SubjectProfile:
      type: object
      properties:
        id:
          type: integer
        subject_id:
          type: integer
        sex:
          type: string
          nullable: true
        birth_year:
          type: integer
          nullable: true
        height_cm:
          type: integer
          nullable: true
        dominant_hand:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SubjectProfileUpdate:
      type: object
      properties:
        sex:
          type: string
        birth_year:
          type: integer
        height_cm:
          type: integer
        dominant_hand:
          type: string
...